+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/cslb/CslbExtractionPipelineBuilderIntegrationSpec.scala
@@ -1,9 +1,11 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.cslb
+
+import better.files.File
+import org.broadinstitute.monster.common.PipelineBuilderSpec
+import org.broadinstitute.monster.dap.common._
+import org.broadinstitute.monster.dap.{CslbExtractionPipeline, common}
 
 import java.time.{LocalDate, LocalTime, OffsetDateTime, ZoneOffset}
-import better.files.File
-import common.{Args, GetRecords, MockOkWrapper, RedcapMsgGenerator, RedcapRequestGeneratorParams}
-import org.broadinstitute.monster.common.PipelineBuilderSpec
 
 class CslbExtractionPipelineBuilderIntegrationSpec extends PipelineBuilderSpec[Args] {
   val outputDir = File.newTemporaryDirectory()
